chore: 기본 provider 컴포넌트를 storybook decorators에 적용
